!! Completed 1.3 refactor and fixed previous errors.
Now is the time to begin refactoring what I have to a way that I like it, separating things into their own functions/structs/classes and respective namespaces.
If possible, I'd like to stabilise the branch a little more before merging to develop.
And comments everywhere need to be done before the merge.